DigitalContentsWebApp/requirements.txt
aioredis==1.3.1
asgiref==3.2.10
asn1crypto==1.4.0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.15.18
botocore==1.18.18
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.8
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.10.1
djangorestframework==3.11.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
incremental==17.5.0
isort==5.6.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
moviepy==1.0.3
msgpack==1.0.0
numpy==1.19.4
Pillow==7.2.0
proglog==0.1.9
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
stripe==2.55.0
toml==0.10.1
tqdm==4.52.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
whitenoise==5.1.0
wrapt==1.12.1
zope.interface==5.1.0
